Switch from GSIs on one table to transactions across tables
The rationale for using GSIs was that it was more extensible--that you
can add GSIs later and they'd be auto-populated. But that's not really
the case...for that to work, the keys for the GSIs need to have
already been written, so if you add a new GSI, you still need to go
add attributes to existing data or they won't show up in the new GSI.

So once I realized that the main benefit of using GSIs was not real,
then there were only drawbacks left--eventual consistency.

With transactions, we get strong consistency since we are not using
GSIs at all, but writing to multiple tables atomically. Additionally,
adding "Txn" datastore support becomes trivial. It will also help
when/if S3 support is added, by minimizing the state space.

Cost-wise, GSIs are treated effectively as separate tables, so there
is no cost difference. And DynamoDB transactions support up to 25
actions per transaction, whereas a table only supports up to 20 GSIs.